reinstate Ember and Handlebars to the layout
This commit reintroduces Ember and Handlebars to the file layout
as they have achieved 1.0 status.  Renamed index.html to
index-backbone.html and introduced a parallel index-ember.html
that uses Handlebars for a simple templating example.

Fixed a bug in main-backbone.js that would have caused a
syntax error in IE's JavaScript interpreter.  The last
member of an object had a trailing comma.